feat: Add feature to scan the QR code from the screen
