+++ lib/widgets/modular_widgets/wallet_connect_widgets/qr_screen_scanner.dart
@@ -1,6 +1,4 @@
 import 'dart:io';
-
-import 'package:bot_toast/bot_toast.dart';
 import 'package:clipboard_watcher/clipboard_watcher.dart';
 import 'package:flutter/cupertino.dart';
 import 'package:flutter/material.dart';
@@ -8,23 +6,25 @@ import 'package:flutter/services.dart';
 import 'package:preference_list/preference_list.dart';
 import 'package:screen_capturer/screen_capturer.dart';
 import 'package:path_provider/path_provider.dart';
+import 'package:zenon_syrius_wallet_flutter/utils/constants.dart';
+import 'package:zenon_syrius_wallet_flutter/widgets/reusable_widgets/buttons/buttons.dart';
 import 'package:zxing2/qrcode.dart';
 import 'package:image/image.dart' as img;
 
 final screenCapturer = ScreenCapturer.instance;
 
-class qr_screen_scanner extends StatefulWidget {
-  const qr_screen_scanner({Key? key}) : super(key: key);
+class QrScreenScanner extends StatefulWidget {
+  const QrScreenScanner({Key? key}) : super(key: key);
 
   @override
-  _qr_screen_scannerState createState() => _qr_screen_scannerState();
+  _QrScreenScannerState createState() => _QrScreenScannerState();
 }
 
-class _qr_screen_scannerState extends State<qr_screen_scanner> with ClipboardListener {
+class _QrScreenScannerState extends State<QrScreenScanner>
+    with ClipboardListener {
   bool _isAccessAllowed = false;
 
   CapturedData? _lastCapturedData;
-  // RecognizeTextResponse? _recognizeTextResponse;
   String? _qrCode;
 
   @override
@@ -42,45 +42,37 @@ class _qr_screen_scannerState extends State<qr_screen_scanner> with ClipboardLis
 
   @override
   void onClipboardChanged() async {
-    // bool hasStrings = await Clipboard.hasStrings();
     ClipboardData? newClipboardData =
-    await Clipboard.getData(Clipboard.kTextPlain);
-    BotToast.showText(text: newClipboardData?.text ?? "");
+        await Clipboard.getData(Clipboard.kTextPlain);
     print(newClipboardData?.text ?? "");
   }
 
   void _handleClickCapture(CaptureMode mode) async {
-    Directory directory = await getApplicationDocumentsDirectory();
-    String imageName =
-        'Screenshot-${DateTime.now().millisecondsSinceEpoch}.png';
-    String imagePath =
-        '${directory.path}/text_recognizer/Screenshots/$imageName';
-    _lastCapturedData = await screenCapturer.capture(
-      mode: mode,
-      imagePath: imagePath,
-      silent: true,
-    );
-    if (_lastCapturedData != null) {
-      var image = img.decodePng(File(imagePath).readAsBytesSync())!;
-
-      LuminanceSource source = RGBLuminanceSource(
-          image.width, image.height, image.getBytes().buffer.asInt32List());
-      var bitmap = BinaryBitmap(HybridBinarizer(source));
-
-      var reader = QRCodeReader();
-      var result = reader.decode(bitmap);
-      _qrCode = reader.decode(bitmap).text;
-      // _recognizeTextResponse = await ocrClient
-      //     .use(kBuiltInOcrEngine)
-      //     .recognizeText(RecognizeTextRequest(
-      //       imagePath: imagePath,
-      //     ));
-      BotToast.showText(text: _qrCode ?? 'null');
-    } else {
-      BotToast.showText(text: 'User canceled capture');
-      print('User canceled capture');
-    }
-    setState(() {});
+    // Directory directory = await getApplicationDocumentsDirectory();
+    // String imageName =
+    //     'Screenshot-${DateTime.now().millisecondsSinceEpoch}.png';
+    // String imagePath =
+    //     '${directory.path}/text_recognizer/Screenshots/$imageName';
+    // _lastCapturedData = await screenCapturer.capture(
+    //   mode: mode,
+    //   imagePath: imagePath,
+    //   silent: true,
+    // );
+    // if (_lastCapturedData != null) {
+    //   var image = img.decodePng(File(imagePath).readAsBytesSync())!;
+    //
+    //   LuminanceSource source = RGBLuminanceSource(
+    //       image.width, image.height, image.getBytes().buffer.asInt32List());
+    //   var bitmap = BinaryBitmap(HybridBinarizer(source));
+    //
+    //   var reader = QRCodeReader();
+    //   var result = reader.decode(bitmap);
+    //   _qrCode = reader.decode(bitmap).text;
+    //   print('QR code: $_qrCode');
+    // } else {
+    //   print('User canceled capture');
+    // }
+    // setState(() {});
   }
 
   Widget _buildBody(BuildContext context) {
@@ -94,8 +86,8 @@ class _qr_screen_scannerState extends State<qr_screen_scanner> with ClipboardLis
                 accessoryView: Text('$_isAccessAllowed'),
                 onTap: () async {
                   bool allowed =
-                  await ScreenCapturer.instance.isAccessAllowed();
-                  BotToast.showText(text: 'allowed: $allowed');
+                      await ScreenCapturer.instance.isAccessAllowed();
+                  print('allowed: $allowed');
                   setState(() {
                     _isAccessAllowed = allowed;
                   });
@@ -106,75 +98,43 @@ class _qr_screen_scannerState extends State<qr_screen_scanner> with ClipboardLis
                 onTap: () async {
                   print('requestingAccess');
                   if (!await ScreenCapturer.instance.isAccessAllowed()) {
-                    BotToast.showText(
-                        text:
-                        'Please allow this application to process QR codes from your screen');
                     await ScreenCapturer.instance.requestAccess();
                   }
                 },
               ),
             ],
           ),
-        Center(
-            child: Column(
-              children: [
-                ElevatedButton(
-                  child: const Text('Start recording clipboard'),
-                  onPressed: () {
-                    clipboardWatcher.start();
-                  },
-                ),
-                Container(
-                  height: 20,
-                ),
-                ElevatedButton(
-                  child: const Text('Stop recording clipboard'),
-                  onPressed: () {
-                    clipboardWatcher.stop();
-                  },
-                ),
-              ],
-            )),
-        PreferenceListSection(
-          title: const Text('METHODS'),
+        Column(
           children: [
-            PreferenceListItem(
-              title: const Text('capture'),
-              accessoryView: Row(children: [
-                CupertinoButton(
-                  child: const Text('region'),
-                  onPressed: () {
-                    _handleClickCapture(CaptureMode.region);
-                  },
-                ),
-                CupertinoButton(
-                  child: const Text('screen'),
-                  onPressed: () {
-                    _handleClickCapture(CaptureMode.screen);
-                  },
-                ),
-                CupertinoButton(
-                  child: const Text('window'),
-                  onPressed: () {
-                    _handleClickCapture(CaptureMode.window);
-                  },
-                ),
-              ]),
+        ElevatedButton(
+          child: const Text('Start recording clipboard'),
+          onPressed: () {
+            clipboardWatcher.start();
+          },
+        ),
+        Container(
+          height: 20,
+        ),
+        ElevatedButton(
+          child: const Text('Stop recording clipboard'),
+          onPressed: () {
+            clipboardWatcher.stop();
+          },
+        ),
+          ],
+        ),
+        Row(
+          children: [
+            MyOutlinedButton(
+              text: 'Scan QR code',
+              onPressed: () {
+                _handleClickCapture(CaptureMode.region);
+              },
+              minimumSize: kLoadingButtonMinSize,
             ),
           ],
         ),
         if (_qrCode != null) Text(_qrCode ?? ''),
-        // if (_recognizeTextResponse != null)
-        //   Text(_recognizeTextResponse?.text ?? ''),
-        if (_lastCapturedData != null && _lastCapturedData?.imagePath != null)
-          Container(
-            margin: const EdgeInsets.only(top: 20),
-            width: 400,
-            height: 400,
-            child: Image.file(
-              File(_lastCapturedData!.imagePath!),
-            ),
-          ),
       ],
     );
   }
